Move worker clients to tachyon.client.worker package
