+++ clients/unshaded/src/main/java/tachyon/client/block/BlockStoreContext.java
@@ -22,13 +22,13 @@ import com.google.common.base.Preconditions;
 import com.google.common.base.Throwables;
 
 import tachyon.client.ClientContext;
+import tachyon.client.worker.WorkerClient;
 import tachyon.exception.ExceptionMessage;
 import tachyon.exception.PreconditionMessage;
 import tachyon.thrift.NetAddress;
 import tachyon.thrift.WorkerInfo;
 import tachyon.util.network.NetworkAddressUtils;
 import tachyon.worker.ClientMetrics;
-import tachyon.worker.WorkerClient;
 
 /**
  * A shared context in each client JVM for common block master client functionality such as a pool
