+++ clients/unshaded/src/main/java/tachyon/client/block/LocalBlockOutStream.java
@@ -27,11 +27,11 @@ import com.google.common.io.Closer;
 
 import tachyon.Constants;
 import tachyon.client.ClientContext;
+import tachyon.client.worker.WorkerClient;
 import tachyon.exception.ExceptionMessage;
 import tachyon.util.io.BufferUtils;
 import tachyon.util.io.FileUtils;
 import tachyon.util.network.NetworkAddressUtils;
-import tachyon.worker.WorkerClient;
 
 /**
  * Provides a streaming API to write to a Tachyon block. This output stream will directly write the
